+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+* Support for updating breakpoints and bookmarks after the buffer rewrite.
+
 ## [0.1.0] - 2025-02-20
 
 ### Added